Initial commit of files
This is a paste of the first mockup I did based on the gist.  I shared
this internally at cPanel and the in-house validation generator was
based on this.  I'm going to deviate from the design used there and take
advantage of features in the latest Angular that were not available to
us.